+++ src/mocks/template.ts
@@ -1,32 +1,48 @@
+import casual from "casual";
+import { Visibility } from "../models/Template";
+
 export const data = [
   {
-    "id": 1,
-    "name": "Test Template 1",
-    "visibility": 'Private',
-    "created": "2024-08-01T13:39:21Z",
-    "modified": "2024-08-01T13:39:21Z",
+    id: casual.integer(1, 9999),
+    name: casual.sentence,
+    // ownerId: 123
+    // affiliationId": "http://example.com/orgs/123",
+    visibility: casual.integer(0, 1) == 1 ? Visibility.Private : Visibility.Public,
+    created: new Date(casual.date('YYYY-MM-DD')).toISOString(),
+    modified: new Date(casual.date('YYYY-MM-DD')).toISOString(),
+    currentVersion: `v${casual.integer(1, 99)}`,
+    isDirty: true,
   },
   {
-    "id": 2,
-    "name": "Test Template 2",
-    "description": "Testing template number 2",
-    "created": "2024-07-30T10:36:56Z",
-    "modified": "2024-08-01T13:39:21Z",
-    "visibility": 'Public',
+    id: casual.integer(1, 9999),
+    name: casual.sentence,
+    // ownerId: 123
+    // affiliationId": "http://example.com/orgs/123",
+    visibility: casual.integer(0, 1) == 1 ? Visibility.Private : Visibility.Public,
+    created: new Date(casual.date('YYYY-MM-DD')).toISOString(),
+    modified: new Date(casual.date('YYYY-MM-DD')).toISOString(),
+    currentVersion: `v${casual.integer(1, 99)}`,
+    isDirty: false,
   },
   {
-    "id": 3,
-    "name": "Test Template 3",
-    "created": "2024-08-01T13:39:21Z",
-    "modified": "2024-08-01T13:39:21Z",
-    "visibility": 'Private',
+    id: casual.integer(1, 9999),
+    name: casual.sentence,
+    // ownerId: 123
+    // affiliationId": "http://example.com/orgs/123",
+    visibility: casual.integer(0, 1) == 1 ? Visibility.Private : Visibility.Public,
+    created: new Date(casual.date('YYYY-MM-DD')).toISOString(),
+    modified: new Date(casual.date('YYYY-MM-DD')).toISOString(),
+    isDirty: true,
   },
   {
-    "id": 4,
-    "name": "Test Template 4",
-    "description": "Testing template number 2",
-    "created": "2024-07-30T10:36:56Z",
-    "modified": "2024-08-01T13:39:21Z",
-    "visibility": 'Public',
+    id: casual.integer(1, 9999),
+    name: casual.sentence,
+    // ownerId: 123
+    // affiliationId": "http://example.com/orgs/123",
+    visibility: casual.integer(0, 1) == 1 ? Visibility.Private : Visibility.Public,
+    created: new Date(casual.date('YYYY-MM-DD')).toISOString(),
+    modified: new Date(casual.date('YYYY-MM-DD')).toISOString(),
+    currentVersion: `v${casual.integer(1, 99)}`,
+    isDirty: false,
   }
 ];