+++ src/models/Collaborator.ts
@@ -1,7 +1,7 @@
 export class Collaborator {
   constructor(
     public email: string,
-    public invitedById: number,
+    public invitedById: number = null,
     public userId: number = null,
     public created: string = new Date().toUTCString(),
   ){}
@@ -11,10 +11,13 @@ export class TemplateCollaborator extends Collaborator {
   constructor(
     public templateId: number,
     public email: string,
-    public invitedById: number,
+    public invitedById: number = null,
     public userId: number = null,
     public created: string = new Date().toUTCString(),
   ){
     super(email, invitedById, userId, created);
+
+console.log(this)
+
   }
 }