+++ src/resolver.ts
@@ -6,6 +6,7 @@ import { orcidScalar } from './resolvers/scalars/orcid';
 import { rorScalar } from './resolvers/scalars/ror';
 
 import { resolvers as affiliationResolvers } from './resolvers/affiliation';
+import { resolvers as collaboratorResolvers } from './resolvers/collaborator';
 import { resolvers as contributorRoleResolvers } from './resolvers/contributorRole';
 import { resolvers as dmspResolvers } from './resolvers/dmsp';
 import { resolvers as templateResolvers } from './resolvers/template';
@@ -20,6 +21,7 @@ const scalarResolvers = {
 export const resolvers: IResolvers = mergeResolvers([
   scalarResolvers,
   affiliationResolvers,
+  collaboratorResolvers,
   dmspResolvers,
   contributorRoleResolvers,
   templateResolvers,