+++ src/schema.ts
@@ -2,6 +2,7 @@ import { mergeTypeDefs } from '@graphql-tools/merge';
 
 import { typeDefs as baseTypeDefs } from './schemas/base';
 import { typeDefs as affiliationTypeDefs } from './schemas/affiliation';
+import { typeDefs as collaboratorTypeDefs } from './schemas/collaborator';
 import { typeDefs as contributorRoleTypeDefs } from './schemas/contributorRole';
 import { typeDefs as dmspTypeDefs } from './schemas/dmsp';
 import { typeDefs as templateTypeDefs } from './schemas/template';
@@ -10,6 +11,7 @@ import { typeDefs as userTypeDefs } from './schemas/user';
 export const typeDefs = mergeTypeDefs([
   baseTypeDefs,
   affiliationTypeDefs,
+  collaboratorTypeDefs,
   contributorRoleTypeDefs,
   dmspTypeDefs,
   templateTypeDefs,