+++ src/services/__tests__/emailService.spec.ts
@@ -1,4 +1,3 @@
-import casual from "casual";
 import { send } from "../emailService";
 
 describe('send', () => {