+++ src/types.ts
@@ -210,10 +210,26 @@ export type Mutation = {
   _empty?: Maybe<Scalars['String']['output']>;
   /** Add a new contributor role (URL and label must be unique!) */
   addContributorRole?: Maybe<ContributorRoleMutationResponse>;
+  /** Create a new Template */
+  addTemplate?: Maybe<Template>;
+  /** Add a collaborator to a Template */
+  addTemplateCollaborator?: Maybe<Scalars['Boolean']['output']>;
+  /** Archive a Template (unpublishes any associated PublishedTemplate */
+  archiveTemplate?: Maybe<Scalars['Boolean']['output']>;
+  /** Create a Template from another PublishedTemplate */
+  copyTemplate?: Maybe<Template>;
+  /** Publish a Template */
+  publishTemplate?: Maybe<PublishedTemplate>;
   /** Delete the contributor role */
   removeContributorRole?: Maybe<ContributorRoleMutationResponse>;
+  /** Remove a TemplateCollaborator from a Template */
+  removeTemplateCollaborator?: Maybe<Scalars['Boolean']['output']>;
+  /** Unpublish the specified PublishedTemplate */
+  unpublishTemplate?: Maybe<Scalars['Boolean']['output']>;
   /** Update the contributor role */
   updateContributorRole?: Maybe<ContributorRoleMutationResponse>;
+  /** Update a Template */
+  updateTemplate?: Maybe<Template>;
 };
 
 
@@ -225,8 +241,47 @@ export type MutationAddContributorRoleArgs = {
 };
 
 
+export type MutationAddTemplateArgs = {
+  name: Scalars['String']['input'];
+};
+
+
+export type MutationAddTemplateCollaboratorArgs = {
+  email: Scalars['String']['input'];
+  templateId: Scalars['Int']['input'];
+};
+
+
+export type MutationArchiveTemplateArgs = {
+  templateId: Scalars['Int']['input'];
+};
+
+
+export type MutationCopyTemplateArgs = {
+  name: Scalars['String']['input'];
+  publishedTemplateId: Scalars['Int']['input'];
+};
+
+
+export type MutationPublishTemplateArgs = {
+  comment?: InputMaybe<Scalars['String']['input']>;
+  templateId: Scalars['Int']['input'];
+};
+
+
 export type MutationRemoveContributorRoleArgs = {
   id: Scalars['ID']['input'];
+};
+
+
+export type MutationRemoveTemplateCollaboratorArgs = {
+  email: Scalars['String']['input'];
+  templateId: Scalars['Int']['input'];
+};
+
+
+export type MutationUnpublishTemplateArgs = {
+  publishedTemplateId: Scalars['Int']['input'];
 };
 
 
@@ -236,6 +291,13 @@ export type MutationUpdateContributorRoleArgs = {
   id: Scalars['ID']['input'];
   label: Scalars['String']['input'];
   url: Scalars['URL']['input'];
+};
+
+
+export type MutationUpdateTemplateArgs = {
+  name: Scalars['String']['input'];
+  templateId: Scalars['Int']['input'];
+  visibility: Visibility;
 };
 
 export type OrganizationIdentifier = {
@@ -264,6 +326,33 @@ export type PrimaryContact = Person & {
   name: Scalars['String']['output'];
 };
 
+/** A snapshot of a Template when it became published. DMPs are created from published templates */
+export type PublishedTemplate = {
+  __typename?: 'PublishedTemplate';
+  /** Whether or not this is the version provided when users create a new DMP */
+  active: Scalars['Boolean']['output'];
+  /** The affiliation that the template belongs to */
+  affiliation: Affiliation;
+  /** A comment/note the user enters when publishing the Template */
+  comment?: Maybe<Scalars['String']['output']>;
+  /** The timestamp when the template was created */
+  created: Scalars['DateTimeISO']['output'];
+  /** The unique identifer for the template */
+  id: Scalars['Int']['output'];
+  /** The name/title of the template */
+  name: Scalars['String']['output'];
+  /** The owner of the Template */
+  owner: User;
+  /** The publisher of the Template */
+  publishedBy: User;
+  /** The template that this published version stems from */
+  template: Template;
+  /** The major.minor semantic version */
+  version: Scalars['String']['output'];
+  /** The template's availability setting: Public is available to everyone, Private only your affiliation */
+  visibility: Visibility;
+};
+
 export type Query = {
   __typename?: 'Query';
   _empty?: Maybe<Scalars['String']['output']>;
@@ -271,6 +360,8 @@ export type Query = {
   affiliation?: Maybe<Affiliation>;
   /** Perform a search for Affiliations matching the specified name */
   affiliations?: Maybe<Array<Maybe<AffiliationSearch>>>;
+  /** Get the DMPTool Best Practice PublishedTemplates */
+  bestPracticeTemplates?: Maybe<Array<Maybe<PublishedTemplate>>>;
   /** Get the contributor role by it's id */
   contributorRoleById?: Maybe<ContributorRole>;
   /** Get the contributor role by it's URL */
@@ -280,6 +371,18 @@ export type Query = {
   /** Get the DMSP by its DMP ID */
   dmspById?: Maybe<SingleDmspResponse>;
   me?: Maybe<User>;
+  /** Get the Templates that belong to the current user's affiliation (user must be an Admin) */
+  myTemplates?: Maybe<Array<Maybe<Template>>>;
+  /** Get the specified PublishedTemplate */
+  publishedTemplate?: Maybe<PublishedTemplate>;
+  /** Search for PublishedTemplates whose name or owning Org's name contains the search term */
+  publishedTemplates?: Maybe<Array<Maybe<PublishedTemplate>>>;
+  /** Get the specified Template (user must be an Admin) */
+  template?: Maybe<Template>;
+  /** Get all of the Users that belong to another affiliation that can edit the Template */
+  templateCollaborators?: Maybe<Array<Maybe<TemplateCollaborator>>>;
+  /** Get all of the PublishedTemplates for the specified Template (a.k. the Template history) */
+  templateVersions?: Maybe<Array<Maybe<PublishedTemplate>>>;
   user?: Maybe<User>;
   users?: Maybe<Array<Maybe<User>>>;
 };
@@ -311,6 +414,31 @@ export type QueryDmspByIdArgs = {
 };
 
 
+export type QueryPublishedTemplateArgs = {
+  publishedTemplateId: Scalars['Int']['input'];
+};
+
+
+export type QueryPublishedTemplatesArgs = {
+  term: Scalars['String']['input'];
+};
+
+
+export type QueryTemplateArgs = {
+  templateId: Scalars['Int']['input'];
+};
+
+
+export type QueryTemplateCollaboratorsArgs = {
+  templateId: Scalars['Int']['input'];
+};
+
+
+export type QueryTemplateVersionsArgs = {
+  templateId: Scalars['Int']['input'];
+};
+
+
 export type QueryUserArgs = {
   userId: Scalars['String']['input'];
 };
@@ -335,6 +463,44 @@ export type SingleDmspResponse = {
   success: Scalars['Boolean']['output'];
 };
 
+/** A Template used to create DMPs */
+export type Template = {
+  __typename?: 'Template';
+  /** The affiliation that the template belongs to */
+  affiliation?: Maybe<Affiliation>;
+  /** The timestamp when the template was created */
+  created: Scalars['DateTimeISO']['output'];
+  /** The current published version */
+  currentVersion?: Maybe<Scalars['String']['output']>;
+  /** The unique identifer for the template */
+  id: Scalars['Int']['output'];
+  /** Whether or not the Template has had any changes since it was last published */
+  isDirty: Scalars['Boolean']['output'];
+  /** The timestamp when the template was modifed */
+  modified: Scalars['DateTimeISO']['output'];
+  /** The name/title of the template */
+  name: Scalars['String']['output'];
+  /** The owner of the Template */
+  owner?: Maybe<User>;
+  /** The template's availability setting: Public is available to everyone, Private only your affiliation */
+  visibility: Visibility;
+};
+
+/** A user that that belongs to a different affiliation that can edit the Template */
+export type TemplateCollaborator = {
+  __typename?: 'TemplateCollaborator';
+  /** The timestamp when the collaborator was added */
+  created: Scalars['DateTimeISO']['output'];
+  /** The collaborator's email */
+  email: Scalars['String']['output'];
+  /** The person who invited the collaborator */
+  invitedBy?: Maybe<User>;
+  /** The template the collaborator may edit */
+  template?: Maybe<Template>;
+  /** The collaborator (if they have an account) */
+  user?: Maybe<User>;
+};
+
 export type User = {
   __typename?: 'User';
   affiliation?: Maybe<Affiliation>;
@@ -352,6 +518,13 @@ export type UserRole =
   | 'Admin'
   | 'Researcher'
   | 'SuperAdmin';
+
+/** Template visibility */
+export type Visibility =
+  /** Visible only to users of your institution */
+  | 'Private'
+  /** Visible to all users */
+  | 'Public';
 
 export type YesNoUnknown =
   | 'no'
@@ -458,14 +631,18 @@ export type ResolversTypes = {
   Person: ResolverTypeWrapper<ResolversInterfaceTypes<ResolversTypes>['Person']>;
   PersonIdentifier: ResolverTypeWrapper<PersonIdentifier>;
   PrimaryContact: ResolverTypeWrapper<PrimaryContact>;
+  PublishedTemplate: ResolverTypeWrapper<PublishedTemplate>;
   Query: ResolverTypeWrapper<{}>;
   RelatedIdentifier: ResolverTypeWrapper<RelatedIdentifier>;
   Ror: ResolverTypeWrapper<Scalars['Ror']['output']>;
   SingleDmspResponse: ResolverTypeWrapper<Omit<SingleDmspResponse, 'dmsp'> & { dmsp?: Maybe<ResolversTypes['Dmsp']> }>;
   String: ResolverTypeWrapper<Scalars['String']['output']>;
+  Template: ResolverTypeWrapper<Template>;
+  TemplateCollaborator: ResolverTypeWrapper<TemplateCollaborator>;
   URL: ResolverTypeWrapper<Scalars['URL']['output']>;
   User: ResolverTypeWrapper<User>;
   UserRole: UserRole;
+  Visibility: Visibility;
   YesNoUnknown: YesNoUnknown;
 };
 
@@ -496,11 +673,14 @@ export type ResolversParentTypes = {
   Person: ResolversInterfaceTypes<ResolversParentTypes>['Person'];
   PersonIdentifier: PersonIdentifier;
   PrimaryContact: PrimaryContact;
+  PublishedTemplate: PublishedTemplate;
   Query: {};
   RelatedIdentifier: RelatedIdentifier;
   Ror: Scalars['Ror']['output'];
   SingleDmspResponse: Omit<SingleDmspResponse, 'dmsp'> & { dmsp?: Maybe<ResolversParentTypes['Dmsp']> };
   String: Scalars['String']['output'];
+  Template: Template;
+  TemplateCollaborator: TemplateCollaborator;
   URL: Scalars['URL']['output'];
   User: User;
 };
@@ -644,8 +824,16 @@ export type IdentifierResolvers<ContextType = MyContext, ParentType extends Reso
 export type MutationResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
   _empty?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   addContributorRole?: Resolver<Maybe<ResolversTypes['ContributorRoleMutationResponse']>, ParentType, ContextType, RequireFields<MutationAddContributorRoleArgs, 'displayOrder' | 'label' | 'url'>>;
+  addTemplate?: Resolver<Maybe<ResolversTypes['Template']>, ParentType, ContextType, RequireFields<MutationAddTemplateArgs, 'name'>>;
+  addTemplateCollaborator?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType, RequireFields<MutationAddTemplateCollaboratorArgs, 'email' | 'templateId'>>;
+  archiveTemplate?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType, RequireFields<MutationArchiveTemplateArgs, 'templateId'>>;
+  copyTemplate?: Resolver<Maybe<ResolversTypes['Template']>, ParentType, ContextType, RequireFields<MutationCopyTemplateArgs, 'name' | 'publishedTemplateId'>>;
+  publishTemplate?: Resolver<Maybe<ResolversTypes['PublishedTemplate']>, ParentType, ContextType, RequireFields<MutationPublishTemplateArgs, 'templateId'>>;
   removeContributorRole?: Resolver<Maybe<ResolversTypes['ContributorRoleMutationResponse']>, ParentType, ContextType, RequireFields<MutationRemoveContributorRoleArgs, 'id'>>;
+  removeTemplateCollaborator?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType, RequireFields<MutationRemoveTemplateCollaboratorArgs, 'email' | 'templateId'>>;
+  unpublishTemplate?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType, RequireFields<MutationUnpublishTemplateArgs, 'publishedTemplateId'>>;
   updateContributorRole?: Resolver<Maybe<ResolversTypes['ContributorRoleMutationResponse']>, ParentType, ContextType, RequireFields<MutationUpdateContributorRoleArgs, 'displayOrder' | 'id' | 'label' | 'url'>>;
+  updateTemplate?: Resolver<Maybe<ResolversTypes['Template']>, ParentType, ContextType, RequireFields<MutationUpdateTemplateArgs, 'name' | 'templateId' | 'visibility'>>;
 };
 
 export interface OrcidScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['Orcid'], any> {
@@ -679,15 +867,37 @@ export type PrimaryContactResolvers<ContextType = MyContext, ParentType extends 
   __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
 };
 
+export type PublishedTemplateResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['PublishedTemplate'] = ResolversParentTypes['PublishedTemplate']> = {
+  active?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
+  affiliation?: Resolver<ResolversTypes['Affiliation'], ParentType, ContextType>;
+  comment?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  created?: Resolver<ResolversTypes['DateTimeISO'], ParentType, ContextType>;
+  id?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
+  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  owner?: Resolver<ResolversTypes['User'], ParentType, ContextType>;
+  publishedBy?: Resolver<ResolversTypes['User'], ParentType, ContextType>;
+  template?: Resolver<ResolversTypes['Template'], ParentType, ContextType>;
+  version?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  visibility?: Resolver<ResolversTypes['Visibility'], ParentType, ContextType>;
+  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+};
+
 export type QueryResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
   _empty?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   affiliation?: Resolver<Maybe<ResolversTypes['Affiliation']>, ParentType, ContextType, RequireFields<QueryAffiliationArgs, 'affiliationId'>>;
   affiliations?: Resolver<Maybe<Array<Maybe<ResolversTypes['AffiliationSearch']>>>, ParentType, ContextType, RequireFields<QueryAffiliationsArgs, 'name'>>;
+  bestPracticeTemplates?: Resolver<Maybe<Array<Maybe<ResolversTypes['PublishedTemplate']>>>, ParentType, ContextType>;
   contributorRoleById?: Resolver<Maybe<ResolversTypes['ContributorRole']>, ParentType, ContextType, RequireFields<QueryContributorRoleByIdArgs, 'contributorRoleId'>>;
   contributorRoleByURL?: Resolver<Maybe<ResolversTypes['ContributorRole']>, ParentType, ContextType, RequireFields<QueryContributorRoleByUrlArgs, 'contributorRoleURL'>>;
   contributorRoles?: Resolver<Maybe<Array<Maybe<ResolversTypes['ContributorRole']>>>, ParentType, ContextType>;
   dmspById?: Resolver<Maybe<ResolversTypes['SingleDmspResponse']>, ParentType, ContextType, RequireFields<QueryDmspByIdArgs, 'dmspId'>>;
   me?: Resolver<Maybe<ResolversTypes['User']>, ParentType, ContextType>;
+  myTemplates?: Resolver<Maybe<Array<Maybe<ResolversTypes['Template']>>>, ParentType, ContextType>;
+  publishedTemplate?: Resolver<Maybe<ResolversTypes['PublishedTemplate']>, ParentType, ContextType, RequireFields<QueryPublishedTemplateArgs, 'publishedTemplateId'>>;
+  publishedTemplates?: Resolver<Maybe<Array<Maybe<ResolversTypes['PublishedTemplate']>>>, ParentType, ContextType, RequireFields<QueryPublishedTemplatesArgs, 'term'>>;
+  template?: Resolver<Maybe<ResolversTypes['Template']>, ParentType, ContextType, RequireFields<QueryTemplateArgs, 'templateId'>>;
+  templateCollaborators?: Resolver<Maybe<Array<Maybe<ResolversTypes['TemplateCollaborator']>>>, ParentType, ContextType, RequireFields<QueryTemplateCollaboratorsArgs, 'templateId'>>;
+  templateVersions?: Resolver<Maybe<Array<Maybe<ResolversTypes['PublishedTemplate']>>>, ParentType, ContextType, RequireFields<QueryTemplateVersionsArgs, 'templateId'>>;
   user?: Resolver<Maybe<ResolversTypes['User']>, ParentType, ContextType, RequireFields<QueryUserArgs, 'userId'>>;
   users?: Resolver<Maybe<Array<Maybe<ResolversTypes['User']>>>, ParentType, ContextType>;
 };
@@ -709,6 +919,28 @@ export type SingleDmspResponseResolvers<ContextType = MyContext, ParentType exte
   dmsp?: Resolver<Maybe<ResolversTypes['Dmsp']>, ParentType, ContextType>;
   message?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
   success?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
+  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+};
+
+export type TemplateResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['Template'] = ResolversParentTypes['Template']> = {
+  affiliation?: Resolver<Maybe<ResolversTypes['Affiliation']>, ParentType, ContextType>;
+  created?: Resolver<ResolversTypes['DateTimeISO'], ParentType, ContextType>;
+  currentVersion?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  id?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
+  isDirty?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
+  modified?: Resolver<ResolversTypes['DateTimeISO'], ParentType, ContextType>;
+  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  owner?: Resolver<Maybe<ResolversTypes['User']>, ParentType, ContextType>;
+  visibility?: Resolver<ResolversTypes['Visibility'], ParentType, ContextType>;
+  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+};
+
+export type TemplateCollaboratorResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['TemplateCollaborator'] = ResolversParentTypes['TemplateCollaborator']> = {
+  created?: Resolver<ResolversTypes['DateTimeISO'], ParentType, ContextType>;
+  email?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  invitedBy?: Resolver<Maybe<ResolversTypes['User']>, ParentType, ContextType>;
+  template?: Resolver<Maybe<ResolversTypes['Template']>, ParentType, ContextType>;
+  user?: Resolver<Maybe<ResolversTypes['User']>, ParentType, ContextType>;
   __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
 };
 
@@ -751,10 +983,13 @@ export type Resolvers<ContextType = MyContext> = {
   Person?: PersonResolvers<ContextType>;
   PersonIdentifier?: PersonIdentifierResolvers<ContextType>;
   PrimaryContact?: PrimaryContactResolvers<ContextType>;
+  PublishedTemplate?: PublishedTemplateResolvers<ContextType>;
   Query?: QueryResolvers<ContextType>;
   RelatedIdentifier?: RelatedIdentifierResolvers<ContextType>;
   Ror?: GraphQLScalarType;
   SingleDmspResponse?: SingleDmspResponseResolvers<ContextType>;
+  Template?: TemplateResolvers<ContextType>;
+  TemplateCollaborator?: TemplateCollaboratorResolvers<ContextType>;
   URL?: GraphQLScalarType;
   User?: UserResolvers<ContextType>;
 };